Moved lpworkbench to be a plugin
