+++ website.py
@@ -7,6 +7,7 @@
 from eventlet.green import os, socket
 import json
 import sys
+import traceback
 
 # Global timeout for sockets in case something leaks
 socket.setdefaulttimeout(900)
@@ -193,62 +194,7 @@ class dataSite():
             'servers':self.bitHopper.pool.get_servers(),
             'user':self.bitHopper.data.get_users()})
         return response
-    
-class lpWorkbench():
-    def __init__(self, bitHopper):
-        self.bitHopper = bitHopper        
-        self.site_names = ['/lpworkbench']
-        index_name = 'lpworkbench.html'
-        try:
-            # determine if application is a script file or frozen exe
-            if hasattr(sys, 'frozen'):
-                application_path = os.path.dirname(sys.executable)
-            elif __file__:
-                application_path = os.path.dirname(__file__)          
-            index = os.path.join(application_path, index_name)
-        except:
-            index = os.path.join(os.curdir, index_name)
-        index_file = open(index, 'r')
-        self.line_string = index_file.read()
-        index_file.close()
-        
-    def handle(self, env, start_response):
-        start_response('200 OK', [('Content-Type', 'text/html')])
 
-        #Handle Possible Post values
-        self.handle_POST(webob.Request(env))
-
-        index_name = 'lpworkbench.html'
-        try:
-            # determine if application is a script file or frozen exe
-            if hasattr(sys, 'frozen'):
-                application_path = os.path.dirname(sys.executable)
-            elif __file__:
-                application_path = os.path.dirname(__file__)          
-            index = os.path.join(application_path, index_name)
-        except:
-            index = os.path.join(os.curdir, index_name)
-        index_file = open(index, 'r')
-        self.line_string = index_file.read()
-        index_file.close()
-        return self.line_string
-    
-    def handle_POST(self, request):
-        for v in request.POST:
-            if "setOwner" in v:
-                try:
-                    blockhash = v.split('-')[1]
-                    block = self.bitHopper.lp.blocks[blockhash]
-                    old_owner = None
-                    if block != None and '_owner' in block:
-                        old_owner = self.bitHopper.lp.blocks[blockhash]['_owner']
-                    new_owner = str(request.POST[v])
-                    self.bitHopper.log_msg("Updating Block Owner " + blockhash + " from " + str(old_owner) + ' to ' + str(new_owner))
-                    self.bitHopper.lp.set_owner(new_owner, blockhash)
-                except Exception, e:
-                    self.bitHopper.log_dbg('Incorrect http post request setOwner: ' + str(v))
-                    traceback.print_exc()
-        
 class lpSite():
 
     def __init__(self, bitHopper):
@@ -258,6 +204,7 @@ class lpSite():
 
     def handle(self, env, start_response):
         return self.bitHopper.work.handle_LP(env, start_response)
+
 
 class nullsite():
     def __init__(self):
@@ -283,7 +230,7 @@ class bitSite():
         self.bitHopper = bitHopper
         self.dynamicSite = dynamicSite(self.bitHopper)
         self.sites = [self, lpSite(self.bitHopper), dynamicSite(self.bitHopper),
-                      dataSite(self.bitHopper), lpWorkbench(self.bitHopper)]
+                      dataSite(self.bitHopper)]
 
     def handle_start(self, env, start_response):
         use_site = nosite()
